+++ api_docs/changelog.md
@@ -20,6 +20,12 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 295**
+
+* [`GET /export/realm/consents`](/api/get-realm-export-consents): Added
+  a new endpoint to fetch the consents of users for their [private data
+  exports](/help/export-your-organization#full-export-with-member-consent).
+
 **Feature level 294**
 
 * [`POST /register`](/api/register-queue): Clients that do not